CR: Rename ExecutorURLClassLoader -> SparkURLClassLoader
